Flush data to output CSV after every iteration
Up to that moment every artifact of osmosis was kept in one
ArrayList in RAM until the very end of the program, when it was
written to output CSV file. For big simulations it was causing
OutOfMemoryError as the list had been growing huge.

Therefor the artifacts will be printed after every iteration and
the ArrayList cleared to avoid the issue.